FIXED Enemy Script
We need to add a collider to the player for the basic enemy script to work.
There needs to be 2 boxColliders in each enemy, one for the hitbox, and one for the player area.

I still need to add a onTriggerStay() function so the enemy doesn't push the player around.