Removed Unnecessary BoxCollider, Possible Bug?
I removed an unnecessary BoxCollider.

There is a possibility that there may be a bug where the enemy is accelerating into the player because I saw the enemy bump and move the player in one of the tests I ran.

Also, there was an issue where an object between the player and the enemy in the outer boxcollieder can cause the enemy to stop, and it only starts again when touching the inner boxcollider/enemy model.