Basic Logic for collecting items added, in addition to quest stuff.
Added some items and scripts for the "No Good Son" Quest.
Many items are placeholders and the scripts are in the stages.

Objects disappear when the player moves over them if they have the PickUpItem script on them. We still need to implement an inventory script that keeps track of items the player picks up, since we are abandoning the idea for an inventory screen.